Temporarily disabling webui-war module from being a build depndency
It looks like I may have accidentally deleted the bower.json file for
webui-war.  It'll be possible to recover it, but I don't have the time
to do that right now and this module is likely on its way out anyhow, so
I'm just taking it out of the build as a temporary patch.