Adding a bunch of new dependencies for PostGIS, GeoServer, and Hibernate Spatial architecture
Also added a project to carry the Director (see GoF Builder Design
Pattern) that uses an Excel Reader and the content of an imported Excel
WiNS spreadsheet to drive use of a WiNS model artifact builder.  There
will be a use case in the middletier backend added to retrieve the input
Excel file, invoke the director, and then shuttle the Builder output
into a persistent store (or a slightly more ephemeral in-memory cache if
the user need not save right away?  Perhaps the browser itself?)